Moved netpol e2e tests up level to e2e/network dir
Moved network_policy.go up a level so the actual e2e tests are in the
e2e/network directory with all other e2e tests and the
e2e/network/netpol now contains just the network_policy helpers as
discussed in https://github.com/kubernetes/kubernetes/issues/114369.
Several minor changes were required within the netpol package to
export helpers used by the actual e2e tests.
