+++ test/e2e/network/netpol/kubemanager.go
@@ -19,6 +19,11 @@ package netpol
 import (
 	"context"
 	"fmt"
+	"net"
+	"strconv"
+	"strings"
+	"time"
+
 	v1 "k8s.io/api/core/v1"
 	networkingv1 "k8s.io/api/networking/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -27,10 +32,6 @@ import (
 	"k8s.io/kubernetes/test/e2e/framework"
 	e2epod "k8s.io/kubernetes/test/e2e/framework/pod"
 	netutils "k8s.io/utils/net"
-	"net"
-	"strconv"
-	"strings"
-	"time"
 )
 
 // defaultPollIntervalSeconds [seconds] is the default value for which the Prober will wait before attempting next attempt.
@@ -53,11 +54,11 @@ func (pod TestPod) PodString() PodString {
 	return NewPodString(pod.Namespace, pod.Name)
 }
 
-// kubeManager provides a convenience interface to kube functionality that we leverage for polling NetworkPolicy connections.
+// KubeManager provides a convenience interface to kube functionality that we leverage for polling NetworkPolicy connections.
 // Its responsibilities are:
 //   - creating resources (pods, deployments, namespaces, services, network policies)
 //   - modifying and cleaning up resources
-type kubeManager struct {
+type KubeManager struct {
 	framework      *framework.Framework
 	clientSet      clientset.Interface
 	namespaceNames []string
@@ -66,9 +67,9 @@ type kubeManager struct {
 	dnsDomain      string
 }
 
-// newKubeManager is a utility function that wraps creation of the kubeManager instance.
-func newKubeManager(framework *framework.Framework, dnsDomain string) *kubeManager {
-	return &kubeManager{
+// NewKubeManager is a utility function that wraps creation of the KubeManager instance.
+func NewKubeManager(framework *framework.Framework, dnsDomain string) *KubeManager {
+	return &KubeManager{
 		framework: framework,
 		clientSet: framework.ClientSet,
 		dnsDomain: dnsDomain,
@@ -76,7 +77,7 @@ func newKubeManager(framework *framework.Framework, dnsDomain string) *kubeManag
 }
 
 // initializeCluster initialized the cluster, creating namespaces pods and services as needed.
-func (k *kubeManager) initializeClusterFromModel(ctx context.Context, model *Model) error {
+func (k *KubeManager) InitializeClusterFromModel(ctx context.Context, model *Model) error {
 	var createdPods []*v1.Pod
 	for _, ns := range model.Namespaces {
 		// no labels needed, we just need the default kubernetes.io/metadata.name label
@@ -126,24 +127,24 @@ func (k *kubeManager) initializeClusterFromModel(ctx context.Context, model *Mod
 	return nil
 }
 
-func (k *kubeManager) AllPods() []TestPod {
+func (k *KubeManager) AllPods() []TestPod {
 	return k.allPods
 }
 
-func (k *kubeManager) AllPodStrings() []PodString {
+func (k *KubeManager) AllPodStrings() []PodString {
 	return k.allPodStrings
 }
 
-func (k *kubeManager) DNSDomain() string {
+func (k *KubeManager) DNSDomain() string {
 	return k.dnsDomain
 }
 
-func (k *kubeManager) NamespaceNames() []string {
+func (k *KubeManager) NamespaceNames() []string {
 	return k.namespaceNames
 }
 
 // getPod gets a pod by namespace and name.
-func (k *kubeManager) getPod(ctx context.Context, ns string, name string) (*v1.Pod, error) {
+func (k *KubeManager) getPod(ctx context.Context, ns string, name string) (*v1.Pod, error) {
 	kubePod, err := k.clientSet.CoreV1().Pods(ns).Get(ctx, name, metav1.GetOptions{})
 	if err != nil {
 		return nil, fmt.Errorf("unable to get pod %s/%s: %w", ns, name, err)
@@ -153,7 +154,7 @@ func (k *kubeManager) getPod(ctx context.Context, ns string, name string) (*v1.P
 
 // probeConnectivity execs into a pod and checks its connectivity to another pod.
 // Implements the Prober interface.
-func (k *kubeManager) probeConnectivity(args *probeConnectivityArgs) (bool, string, error) {
+func (k *KubeManager) probeConnectivity(args *probeConnectivityArgs) (bool, string, error) {
 	port := strconv.Itoa(args.toPort)
 	if args.addrTo == "" {
 		return false, "no IP provided", fmt.Errorf("empty addrTo field")
@@ -237,7 +238,7 @@ func (k *kubeManager) probeConnectivity(args *probeConnectivityArgs) (bool, stri
 }
 
 // executeRemoteCommand executes a remote shell command on the given pod.
-func (k *kubeManager) executeRemoteCommand(namespace string, pod string, containerName string, command []string) (string, string, error) {
+func (k *KubeManager) executeRemoteCommand(namespace string, pod string, containerName string, command []string) (string, string, error) {
 	return e2epod.ExecWithOptions(k.framework, e2epod.ExecOptions{
 		Command:            command,
 		Namespace:          namespace,
@@ -251,7 +252,7 @@ func (k *kubeManager) executeRemoteCommand(namespace string, pod string, contain
 }
 
 // createService is a convenience function for service setup.
-func (k *kubeManager) createService(ctx context.Context, service *v1.Service) (*v1.Service, error) {
+func (k *KubeManager) createService(ctx context.Context, service *v1.Service) (*v1.Service, error) {
 	ns := service.Namespace
 	name := service.Name
 
@@ -263,7 +264,7 @@ func (k *kubeManager) createService(ctx context.Context, service *v1.Service) (*
 }
 
 // createPod is a convenience function for pod setup.
-func (k *kubeManager) createPod(ctx context.Context, pod *v1.Pod) (*v1.Pod, error) {
+func (k *KubeManager) createPod(ctx context.Context, pod *v1.Pod) (*v1.Pod, error) {
 	ns := pod.Namespace
 	framework.Logf("creating pod %s/%s", ns, pod.Name)
 
@@ -275,7 +276,7 @@ func (k *kubeManager) createPod(ctx context.Context, pod *v1.Pod) (*v1.Pod, erro
 }
 
 // cleanNetworkPolicies is a convenience function for deleting network policies before startup of any new test.
-func (k *kubeManager) cleanNetworkPolicies(ctx context.Context) error {
+func (k *KubeManager) CleanNetworkPolicies(ctx context.Context) error {
 	for _, ns := range k.namespaceNames {
 		framework.Logf("deleting policies in %s ..........", ns)
 		l, err := k.clientSet.NetworkingV1().NetworkPolicies(ns).List(ctx, metav1.ListOptions{})
@@ -294,7 +295,7 @@ func (k *kubeManager) cleanNetworkPolicies(ctx context.Context) error {
 }
 
 // createNetworkPolicy is a convenience function for creating network policies.
-func (k *kubeManager) createNetworkPolicy(ctx context.Context, ns string, netpol *networkingv1.NetworkPolicy) (*networkingv1.NetworkPolicy, error) {
+func (k *KubeManager) createNetworkPolicy(ctx context.Context, ns string, netpol *networkingv1.NetworkPolicy) (*networkingv1.NetworkPolicy, error) {
 	framework.Logf("creating network policy %s/%s", ns, netpol.Name)
 	netpol.ObjectMeta.Namespace = ns
 	np, err := k.clientSet.NetworkingV1().NetworkPolicies(ns).Create(ctx, netpol, metav1.CreateOptions{})
@@ -305,7 +306,7 @@ func (k *kubeManager) createNetworkPolicy(ctx context.Context, ns string, netpol
 }
 
 // updateNetworkPolicy is a convenience function for updating network policies.
-func (k *kubeManager) updateNetworkPolicy(ctx context.Context, ns string, netpol *networkingv1.NetworkPolicy) (*networkingv1.NetworkPolicy, error) {
+func (k *KubeManager) updateNetworkPolicy(ctx context.Context, ns string, netpol *networkingv1.NetworkPolicy) (*networkingv1.NetworkPolicy, error) {
 	framework.Logf("updating network policy %s/%s", ns, netpol.Name)
 	netpol.ObjectMeta.Namespace = ns
 	np, err := k.clientSet.NetworkingV1().NetworkPolicies(ns).Update(ctx, netpol, metav1.UpdateOptions{})
@@ -316,7 +317,7 @@ func (k *kubeManager) updateNetworkPolicy(ctx context.Context, ns string, netpol
 }
 
 // getNamespace gets a namespace object from kubernetes.
-func (k *kubeManager) getNamespace(ctx context.Context, ns string) (*v1.Namespace, error) {
+func (k *KubeManager) getNamespace(ctx context.Context, ns string) (*v1.Namespace, error) {
 	selectedNameSpace, err := k.clientSet.CoreV1().Namespaces().Get(ctx, ns, metav1.GetOptions{})
 	if err != nil {
 		return nil, fmt.Errorf("unable to get namespace %s: %w", ns, err)
